backend/requirements.txt
Django==3.2.16
django-filter==23.5
djangorestframework==3.12.4
Pillow==9.3.0
psycopg2-binary==2.9.3 
djoser==2.1.0 
gunicorn==20.1.0
drf-extra-fields
